* master: (28 commits)
  .net core 3.0 in maintenance mode and end of support will be in march 2020 (#424)
  export to zipkin is not required to run this app (#423)
  suggesting to add Mike as approver (#422)
  Support using custom text format (#417)
  Fix W3C trace-context test suite tests for tracestate (#418)
  Make BlankSpan, SpanContext.Blank internal (#419)
  Cleanup: Remove SetAttribute(string) and AddEvent(name, attributes) overloads on Span (#412)
  Fix the Span Event name to Jaeger Log field name translation (#415)
  parent d49ce958a33266caa2c1f569f8d9b20597e28a92 (#411)
  Allow sring, long, double and bool resource attributes (#414)
  Move test infra to .NET Core 3.0 (#403)
  Follow capitalization conventions for .NET in metrics (#408)
  Thread safe implementation of Counter aggregator and tests. (#406)
  Make ITracer an abstract class and rename to tracer (#402)
  Move LabelSet implementation from API to SDK (#400)
  Jaeger Span conversion missing span.kind (#401)
  Adding New Relic Exporter links (#396)
  Do not throw errors from DistributedContext in runtime (#388)
  NoOp meters (#393)
  Remove old ID format from Application Insights exporter (#385)
  ...

# Conflicts:
#	samples/Exporters/TestPrometheus.cs
#	src/OpenTelemetry.Exporter.Prometheus/OpenTelemetry.Exporter.Prometheus.csproj
#	src/OpenTelemetry.Exporter.Prometheus/PrometheusExporter.cs
#	src/OpenTelemetry.Exporter.Prometheus/PrometheusExporterHttpServer.cs
#	test/OpenTelemetry.Exporter.Prometheus.Tests/PrometheusExporterTests.cs

+++ test/OpenTelemetry.Shims.OpenTracing.Tests/SpanShimTests.cs
@@ -231,7 +231,7 @@ namespace OpenTelemetry.Shims.OpenTracing.Tests
 
             Assert.Single(spanMock.Attributes);
             Assert.Equal("foo", spanMock.Attributes.First().Key);
-            Assert.Equal(1, (long)spanMock.Attributes.First().Value);
+            Assert.Equal(1L, spanMock.Attributes.First().Value);
         }
 
         [Fact]
@@ -297,7 +297,7 @@ namespace OpenTelemetry.Shims.OpenTracing.Tests
 
             Assert.Single(spanMock.Attributes);
             Assert.Equal("foo", spanMock.Attributes.First().Key);
-            Assert.Equal(1, (long)spanMock.Attributes.First().Value);
+            Assert.Equal(1L, spanMock.Attributes.First().Value);
         }
 
         [Fact]
@@ -314,7 +314,7 @@ namespace OpenTelemetry.Shims.OpenTracing.Tests
             Assert.Equal(2, spanMock.Attributes.Count);
 
             Assert.Equal("foo", spanMock.Attributes.First().Key);
-            Assert.Equal(1, (long)spanMock.Attributes.First().Value);
+            Assert.Equal(1L, spanMock.Attributes.First().Value);
 
             Assert.Equal("bar", spanMock.Attributes.Last().Key);
             Assert.Equal("baz", (string)spanMock.Attributes.Last().Value);
